fix intermittent nvblox mesh dropout
- launch realsense node together with nvblox in a container
- add launch argument `standalone` to run nvblox separately

Problem Description:
After an arbitrary amount of meshes, the nvblox node would stop publishing meshes. The synchronizer callback for the depth camera images + info would stop firing. In order to reduce potential jitter and latency we launch the realsense driver in a container together with nvblox. Seems to fix this issue.